ethereum_gravitynft, gravity utils cleanup
