+++ orchestrator/ethereum_gravitynft/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "ethereum_gravitynft"
+version = "0.1.0"
+authors = ["Jakub Kobeldys <jakub@kobeldys.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+gravity_utils = {path = "../gravity_utils"}
+
+deep_space = {workspace = true}
+clarity = {workspace = true}
+web30 = {workspace = true}
+num256 = {workspace = true}
+log = "0.4"
+sha3 = "0.10"
+
+[dev_dependencies]
+rand = "0.8"
+actix = "0.13"
