+++ orchestrator/ethereum_gravitynft/src/lib.rs
@@ -1,0 +1,7 @@
+//! This crate contains various components and utilities for interacting with the GravityErc721 Ethereum contract.
+
+#[macro_use]
+extern crate log;
+
+pub mod send_erc721_to_cosmos;
+pub mod utils;
